requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-JWT==0.3.2
python-decouple==3.4
